refactor: 메시지 전송 로직 분리
