+++ src/pages/model/Message.ts
@@ -1,0 +1,4 @@
+export default interface Message {
+  text: string[];
+  sent: boolean;
+}
